I moved the logic related to the company dashboard to accounts application
