+++ base/accounts/templates/applications_list.html
