+++ base/accounts/templates/company_dashboard.html
@@ -7,7 +7,7 @@
 
 
 <div class="container">
-    <a href="{% url 'offers:create' %}">Add new offer</a>
+    <a href="{% url 'accounts:create' %}">Add new offer</a>
     <a href="{% url 'accounts:update_profile' user.id %}">Edit company information</a>
 </div>
 
@@ -22,9 +22,9 @@
                 </div>
                   <div class="d-flex justify-content-between">
                     <p class="card-text"><small class="text-body-secondary">{{object.date_created}}</small></p>
-                    <a class="card-text" href="{% url 'offers:applications' object.id %}">Browse candidates</a>
-                    <a class="card-text" href="{% url 'offers:edit' object.id %}">Edit</a>
-                      <form method="POST" action="{% url 'offers:delete' object.id %}">
+                    <a class="card-text" href="{% url 'accounts:applications' object.id %}">Browse candidates</a>
+                    <a class="card-text" href="{% url 'accounts:edit' object.id %}">Edit</a>
+                      <form method="POST" action="{% url 'accounts:delete' object.id %}">
                           {% csrf_token %}
                           <button type="submit" class="btn btn-danger">DELETE</button>
                       </form>
