+++ base/accounts/urls.py
@@ -10,4 +10,8 @@ urlpatterns = [
     path('dashboard/', views.CompanyDashboard.as_view(), name='dashboard'),
     path('profile/update/<int:pk>/', views.ProfileUpdateView.as_view(), name='update_profile'),
     path('profile/', views.UserProfileView.as_view(), name='user_profile'),
+    path('offer/edit/<int:pk>/', views.OfferUpdateView.as_view(), name='edit'),
+    path('offer/create/', views.OfferCreateView.as_view(), name='create'),
+    path('offer/delete/<int:pk>/', views.OfferDeleteView.as_view(), name='delete'),
+    path('applications/<int:offer_id>/', views.ApplicationsListView.as_view(), name='applications'),
 ]
