+++ base/accounts/views.py
@@ -7,9 +7,11 @@ from django.contrib.auth.views import LogoutView
 from django import views
 from django.utils.decorators import method_decorator
 from .auth import company_required, user_required
-from offers.models import Offer, Application
-from django.views.generic import UpdateView
 from .models import CustomUser
+from django.views.generic import ListView, CreateView, UpdateView, DeleteView
+from django.db.models import QuerySet
+from offers.models import Offer, Application
+from typing import Any , Dict
 
 
 class RegisterUserView(FormView):
@@ -79,3 +81,60 @@ class UserProfileView(views.View):
         return render(request,
                       'user_profile.html',
                       )
+
+
+class ApplicationsListView(ListView):
+    model = Application
+    paginate_by = 20
+    template_name = 'applications_list.html'
+
+    def get_queryset(self) -> QuerySet[Any]:
+        queryset = super().get_queryset()
+        return queryset
+
+    def get_context_data(self, **kwargs: Any) -> Dict[str, Any]:
+        context = super().get_context_data(**kwargs)
+        return context
+
+
+class OfferDeleteView(DeleteView):
+    model = Offer
+    success_url = "/"
+
+    def get_queryset(self):
+        queryset = super().get_queryset()
+        queryset = queryset.filter(company=self.request.user.id)
+        return queryset
+
+
+class OfferCreateView(CreateView):
+    model = Offer
+    fields = [
+        'name', 'description', 'level', 'requirements', 'localization',
+        'contract', 'position', 'salary_from', 'salary_to', 'remote'
+    ]
+    template_name = 'offer_create.html'
+    success_url = "/"
+
+    def form_valid(self, form):
+        form.instance.company = self.request.user
+        return super().form_valid(form)
+
+    def get_queryset(self):
+        queryset = super().get_queryset()
+        queryset = queryset.filter(company=self.request.user.id)
+        return queryset
+
+
+class OfferUpdateView(UpdateView):
+    model = Offer
+    fields = [
+        'name', 'description', 'level', 'localization', 'contract', 'position', 'salary_from', 'salary_to', 'remote'
+    ]
+    template_name = 'offer_update.html'
+    success_url = "/"
+
+    def get_queryset(self):
+        queryset = super().get_queryset()
+        queryset = queryset.filter(company=self.request.user.id)
+        return queryset
